src/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "celery[rabbitmq]>=5.3",
    "boto3>=1.34",
    "structlog>=24.1",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "sentence-transformers>=3.0",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]
